requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
requests==2.31.0
python-dotenv==1.0.0
google-auth==2.29.0
google-api-python-client==2.127.0
